rename and extract ontology inferencer
